hopefully fixed error handler

+++ utils/views.py
@@ -89,7 +89,7 @@ class ConfirmView(discord.ui.View):
 
 
 class DismissView(discord.ui.View):
-    def __init__(self, error_id: int, author: discord.Member | discord.User, bot, developer_message: discord.WebhookMessage):
+    def __init__(self, error_id: int, author: discord.Member | discord.User, bot, developer_message: discord.WebhookMessage | None):
         super().__init__(timeout=None)
         self.author = author
         self.error_id = error_id
@@ -112,7 +112,8 @@ class DismissView(discord.ui.View):
                 f"Error {self.error_id} has been dismissed!",
                 ephemeral=True,
             )
-            await self.developer_message.delete()
+            if self.developer_message is not None:
+                await self.developer_message.delete()
             
             self.stop()
             return
